+++ Krakoa/Source Files/Platform/OpenGL/OpenGLRendererAPI.hpp
@@ -12,7 +12,7 @@ namespace krakoa::graphics
 		void DrawIndexed(const iVertexArray& vertexArray) override;
 		void SetClearColour(const kmaths::Vector4f& colour) override;
 
-		void SetViewport(const float x, const float y, const float width, const float height) const noexcept override;
+		void SetViewport(const int x, const int y, const int width, const int height) const noexcept override;
 
 	private:
 		void OutputRenderingArchitecture() override;
